Separates resizing operations & asset delivery into separate mechanisms
This allows one to be a public facing route and the other to hide behind a VPC and run in the background.

Delivery becomes quick and straightforward, while asset resizing is buffered through a PubSub request
