+++ errors.go
@@ -1,8 +1,15 @@
-package main
+package asset_delivery
 
 import (
+	"errors"
 	"fmt"
+	"log"
 	"net/http"
+)
+
+var (
+	ErrFileNotHandled = errors.New("file type not handled")
+	ErrInvalidBounds  = errors.New("invalid image bounds")
 )
 
 type RootError interface {
@@ -47,4 +54,17 @@ func (err *ParamError) Error() string {
 
 func (err *ParamError) Root() error {
 	return err.RootError
+}
+
+func WriteError(w http.ResponseWriter, err error) {
+	status := http.StatusInternalServerError
+	if v, ok := err.(HTTPError); ok {
+		status = v.Status()
+	}
+	w.WriteHeader(status)
+	w.Write([]byte(err.Error()))
+
+	if v, ok := err.(RootError); ok && v.Root() != nil {
+		log.Println(v.Root())
+	}
 }
