+++ file-system-gcloud.go
@@ -1,4 +1,4 @@
-package main
+package asset_delivery
 
 import (
 	"context"
