+++ go.mod
@@ -3,6 +3,8 @@ module github.com/monstercat/asset-delivery
 go 1.16
 
 require (
+	cloud.google.com/go/kms v1.1.0 // indirect
+	cloud.google.com/go/pubsub v1.3.1
 	cloud.google.com/go/storage v1.18.2
 	github.com/chai2010/webp v1.1.0
 	github.com/disintegration/imaging v1.6.2
